backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
yt-dlp==2024.3.10
python-multipart==0.0.9
httpx==0.26.0
openai==1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
beautifulsoup4==4.12.3
